refactor(remove rateQuote from redux state)
