+++ packages/blockchain-wallet-v4-frontend/src/modals/CoinifyExchangeData/Confirm/index.js
@@ -43,7 +43,6 @@ class ConfirmContainer extends Component {
           value={value}
           ui={ui}
           medium={medium}
-          rateQuote={this.props.rateQuote}
           onSubmit={this.onSubmit}
           editingAmount={editingAmount}
           toggleEdit={() => this.props.updateUI({ editing: !this.props.ui.editing })}
@@ -57,7 +56,6 @@ class ConfirmContainer extends Component {
 
 const mapStateToProps = (state) => ({
   data: getData(state),
-  rateQuote: selectors.core.data.coinify.getRateQuote(state),
   medium: path(['coinify', 'medium'], state),
   editingAmount: formValueSelector('coinifyConfirm')(state, 'amount')
 })
