+++ packages/blockchain-wallet-v4-frontend/src/modals/CoinifyExchangeData/Confirm/template.js
@@ -6,6 +6,7 @@ import { spacing } from 'services/StyleService'
 
 import { Button, HeartbeatLoader, Text, Link } from 'blockchain-info-components'
 import { Form, ColLeft, InputWrapper, PartnerHeader, PartnerSubHeader, ColRight } from 'components/BuySell/Signup'
+import { getRateFromQuote } from 'services/CoinifyService'
 
 const SummaryWrapper = styled.div`
   display: flex;
@@ -40,7 +41,7 @@ const Unit = styled.div`
 `
 
 const Confirm = (props) => {
-  const { ui, value, onSubmit, busy, invalid, submitting, medium, toggleEdit, editingAmount, rateQuote } = props
+  const { ui, value, onSubmit, busy, invalid, submitting, medium, toggleEdit, editingAmount } = props
   const { quote, mediums, limits } = value
   const { quoteAmount, baseAmount, baseCurrency } = quote
   const { total, fee } = mediums[medium]
@@ -74,7 +75,7 @@ const Confirm = (props) => {
           </PartnerSubHeader>
         </InputWrapper>
         <RateContainer style={spacing('mb-10')}>
-          { rateQuote ? <Text size='12px' weight={300}>1 BTC = {rateQuote.quoteAmount} {rateQuote.quoteCurrency}</Text> : <span>Loading...</span> }
+          { quote ? <Text size='12px' weight={300}>{getRateFromQuote(quote)}</Text> : <span>Loading...</span> }
         </RateContainer>
         <SummaryWrapper>
           <SummaryRow borderBottom>
