+++ packages/blockchain-wallet-v4/src/redux/data/coinify/reducers.js
@@ -8,7 +8,6 @@ const INITIAL_STATE = {
   trades: Remote.NotAsked,
   profile: Remote.NotAsked,
   mediums: Remote.NotAsked,
-  rateQuote: Remote.NotAsked,
   kycs: Remote.NotAsked,
   nextAddress: null
 }
@@ -34,15 +33,6 @@ const coinifyReducer = (state = INITIAL_STATE, action) => {
     }
     case AT.COINIFY_FETCH_QUOTE_FAILURE: {
       return assoc('quote', Remote.Failure(payload), state)
-    }
-    case AT.COINIFY_FETCH_RATE_QUOTE_LOADING: {
-      return assoc('rateQuote', Remote.Loading, state)
-    }
-    case AT.COINIFY_FETCH_RATE_QUOTE_SUCCESS: {
-      return assoc('rateQuote', Remote.Success(payload), state)
-    }
-    case AT.COINIFY_FETCH_RATE_QUOTE_FAILURE: {
-      return assoc('rateQuote', Remote.Failure(payload), state)
     }
     case AT.COINIFY_FETCH_TRADES_LOADING: {
       return assoc('trades', Remote.Loading, state)
